+++ src/main/java/codesquad/server/Server.java
@@ -1,4 +1,4 @@
-package codesquad;
+package codesquad.server;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
